samples/python/agents/google_adk_with_agent_auth/pyproject.toml
[project]
name = "a2a-sample-agent-adk-with-agent-auth"
version = "0.1.0"
description = "Sample Google ADK-based agent that requires authentication."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-samples",
    "click>=8.1.8",
    "google-adk>=0.0.3",
    "google-api-python-client>=2.167.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.9.0",
    "itsdangerous>=2.2.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
a2a-samples = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
